Compatibility with Foundry 13
* Updated hooks to the combat tracker to enable the interface to be opened normally again.
* Removed patches.js as it is no longer necessary.
